Refactor Supabase client implementation; update file paths and add dotenv support
