+++ .gitignore
@@ -1,3 +1,5 @@
+.env
+
 # Logs
 logs
 *.log
