+++ app-electron/src/backend/db/supabaseClient.js
@@ -1,7 +1,7 @@
 const {createClient} = require('@supabase/supabase-js');
 
-const SUPABASE_URL = 'https://jrgubxunsvsnlwkvxfen.supabase.co';
-const SUPABASE_ANON_KEY = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImpyZ3VieHVuc3Zzbmx3a3Z4ZmVuIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDkyNjU5MTYsImV4cCI6MjA2NDg0MTkxNn0.KhsYxvf30JJi194Wi_AlujqfiLZvCcZajdVyRni6ulQ';
+const SUPABASE_URL = process.env.SUPABASE_URL;
+const SUPABASE_ANON_KEY = process.env.SUPABASE_ANON_KEY;
 
 if (!SUPABASE_URL || !SUPABASE_ANON_KEY) {
   throw new Error('Faltan variables de entorno para Supabase. Asegúrate de definir SUPABASE_URL y SUPABASE_ANON_KEY.');
