+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.50.0",
     "@tailwindcss/cli": "^4.1.8",
+    "dotenv": "^16.5.0",
     "mysql2": "^3.14.1",
     "tailwindcss": "^4.1.8"
   }
